Compute base quantity with max buy power

+++ cpp/sandbox/src/alpaca/tao_bot/fetch_account_balance.cpp
@@ -4,7 +4,7 @@
 #include "deps.cpp"           // json, nlohmann
 #include "lib/utils/json.cpp" // ::utils::json
 #include "tao_bot.h"          // Alpaca::TaoBot
-#include <string>             // std::stod, std::stoi, std::string
+#include <string>             // std::stod, std::string
 
 Alpaca::TaoBot::account_balance_t Alpaca::TaoBot::fetch_account_balance() {
   try {
@@ -19,7 +19,7 @@ Alpaca::TaoBot::account_balance_t Alpaca::TaoBot::fetch_account_balance() {
     return {
         .balance = std::stod(balance),
         .margin_buying_power = std::stod(margin_buying_power),
-        .margin_multiplier = std::stoi(margin_multiplier),
+        .margin_multiplier = std::stod(margin_multiplier),
     };
   } catch (nlohmann::detail::type_error &) {
     return fetch_account_balance();
